requirements-dev.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
fonttools==4.55.0
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.0
h5py==2.10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
joblib==1.4.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.3.4
numpy==1.18.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==1.0.5
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
six==1.16.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==2.4.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2
awsiotsdk==1.22.0